+++ client-v2/src/test/java/com/clickhouse/client/insert/InsertTests.java
@@ -11,6 +11,7 @@ import com.clickhouse.client.api.Client;
 import com.clickhouse.client.api.enums.Protocol;
 import com.clickhouse.client.api.insert.InsertResponse;
 import com.clickhouse.client.api.insert.InsertSettings;
+import com.clickhouse.client.api.metadata.TableSchema;
 import com.clickhouse.client.api.metrics.ClientMetrics;
 import com.clickhouse.client.api.metrics.OperationMetrics;
 import com.clickhouse.client.api.metrics.ServerMetrics;
@@ -18,6 +19,7 @@ import org.testcontainers.shaded.org.apache.commons.lang3.RandomStringUtils;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
@@ -31,7 +33,7 @@ public class InsertTests extends BaseIntegrationTest {
     private InsertSettings settings;
 
     @BeforeMethod(groups = { "integration" }, enabled = true)
-    public void setUp() {
+    public void setUp() throws IOException {
         ClickHouseNode node = getServer(ClickHouseProtocol.HTTP);
         client = new Client.Builder()
                 .addEndpoint(Protocol.HTTP, node.getHost(), node.getPort(), false)
@@ -51,19 +53,29 @@ public class InsertTests extends BaseIntegrationTest {
         }
     }
 
+    private void dropTable(String tableName) throws ClickHouseException {
+        try (ClickHouseClient client = ClickHouseClient.builder().config(new ClickHouseConfig())
+                .nodeSelector(ClickHouseNodeSelector.of(ClickHouseProtocol.HTTP))
+                .build()) {
+            String tableQuery = "DROP TABLE IF EXISTS " + tableName;
+            client.read(getServer(ClickHouseProtocol.HTTP)).query(tableQuery).executeAndWait().close();
+        }
+    }
+
     @Test(groups = { "integration" }, enabled = true)
     public void insertSimplePOJOs() throws Exception {
         String tableName = "simple_pojo_table";
         String createSQL = SamplePOJO.generateTableCreateSQL(tableName);
+        String uuid = UUID.randomUUID().toString();
         System.out.println(createSQL);
         createTable(createSQL);
-
-        client.register(SamplePOJO.class, SamplePOJO.generateTableSchema(tableName));
+        client.register(SamplePOJO.class, client.getTableSchema(tableName, "default"));
         List<Object> simplePOJOs = new ArrayList<>();
 
         for (int i = 0; i < 1000; i++) {
             simplePOJOs.add(new SamplePOJO());
         }
+        settings.setQueryId(uuid);
         InsertResponse response = client.insert(tableName, simplePOJOs, settings).get(30, TimeUnit.SECONDS);
 
         OperationMetrics metrics = response.getMetrics();
@@ -71,5 +83,8 @@ public class InsertTests extends BaseIntegrationTest {
         assertEquals(simplePOJOs.size(), response.getWrittenRows());
         assertTrue(metrics.getMetric(ClientMetrics.OP_DURATION).getLong() > 0);
         assertTrue(metrics.getMetric(ClientMetrics.OP_SERIALIZATION).getLong() > 0);
+        assertEquals(metrics.getQueryId(), uuid);
+        assertEquals(response.getQueryId(), uuid);
+        dropTable(tableName);
     }
 }